--- src/main/java/ru/andnovikov/ijustwannarun/IJustWannaRunApplication.java
@@ -1,13 +1,0 @@
-package ru.andnovikov.ijustwannarun;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class IJustWannaRunApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(IJustWannaRunApplication.class, args);
-	}
-
-}
