--- src/main/java/ru/andnovikov/ijustwannarun/controller/UserRegistrationController.java
@@ -1,11 +1,0 @@
-package ru.andnovikov.ijustwannarun.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.ControllerAdvice;
-
-@Controller
-public class UserRegistrationController {
-
-
-
-}
