--- src/main/java/ru/andnovikov/ijustwannarun/controller/WelcomeController.java
@@ -1,14 +1,0 @@
-package ru.andnovikov.ijustwannarun.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-public class WelcomeController {
-
-    @GetMapping("/")
-    public String welcome(){
-        return "events";
-    }
-
-}
