--- src/main/java/ru/andnovikov/ijustwannarun/domain/enumeration/EventKind.java
@@ -1,5 +1,0 @@
-package ru.andnovikov.ijustwannarun.domain.enumeration;
-
-public enum EventKind {
-    RUNNING, SWIMMING, TRIATHLON, BIKE, SKIING;
-}
