--- src/main/java/ru/andnovikov/ijustwannarun/domain/enumeration/EventStatus.java
@@ -1,5 +1,0 @@
-package ru.andnovikov.ijustwannarun.domain.enumeration;
-
-public enum EventStatus {
-    PLAN, OPEN, CLOSED, COMPLETE
-}
