--- src/main/java/ru/andnovikov/ijustwannarun/service/EventServiceImpl.java
@@ -1,74 +1,0 @@
-package ru.andnovikov.ijustwannarun.service;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
-import org.springframework.stereotype.Service;
-import ru.andnovikov.ijustwannarun.domain.Event;
-import ru.andnovikov.ijustwannarun.repository.EventRepository;
-
-import java.util.Optional;
-
-/**
- * Service Implementation for managing {@link Event}.
- */
-@Service
-public class EventServiceImpl implements EventService {
-
-    private final Logger log = LoggerFactory.getLogger(EventServiceImpl.class);
-
-    private final EventRepository eventRepository;
-
-    public EventServiceImpl(EventRepository eventRepository) {
-        this.eventRepository = eventRepository;
-    }
-
-    /**
-     * Save a events.
-     *
-     * @param event the entity to save.
-     * @return the persisted entity.
-     */
-    @Override
-    public Event save(Event event) {
-        log.debug("Request to save Event : {}", event);
-        return eventRepository.save(event);
-    }
-
-    /**
-     * Get all the event.
-     *
-     * @param pageable the pagination information.
-     * @return the list of entities.
-     */
-    @Override
-    public Page<Event> findAll(Pageable pageable) {
-        log.debug("Request to get all Event");
-        return eventRepository.findAll(pageable);
-    }
-
-
-    /**
-     * Get one event by id.
-     *
-     * @param id the id of the entity.
-     * @return the entity.
-     */
-    @Override
-    public Optional<Event> findOne(String id) {
-        log.debug("Request to get Event : {}", id);
-        return eventRepository.findById(id);
-    }
-
-    /**
-     * Delete the event by id.
-     *
-     * @param id the id of the entity.
-     */
-    @Override
-    public void delete(String id) {
-        log.debug("Request to delete Event : {}", id);
-        eventRepository.deleteById(id);
-    }
-}
