+++ src/main/java/ru/andnovikov/sportnow/IJustWannaRunApplication.java
@@ -1,0 +1,13 @@
+package ru.andnovikov.sportnow;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class IJustWannaRunApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(IJustWannaRunApplication.class, args);
+	}
+
+}
