+++ src/main/java/ru/andnovikov/sportnow/controller/WelcomeController.java
@@ -1,0 +1,14 @@
+package ru.andnovikov.sportnow.controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class WelcomeController {
+
+    @GetMapping("/")
+    public String welcome(){
+        return "events";
+    }
+
+}
