+++ src/main/java/ru/andnovikov/sportnow/domain/UserRegistration.java
@@ -1,0 +1,53 @@
+package ru.andnovikov.sportnow.domain;
+
+import org.springframework.data.annotation.Id;
+import org.springframework.data.mongodb.core.mapping.DBRef;
+import org.springframework.data.mongodb.core.mapping.Document;
+import org.springframework.data.mongodb.core.mapping.Field;
+
+import java.util.Date;
+
+@Document(collection = "user_registration")
+public class UserRegistration {
+
+    @Id
+    private String id;
+    @DBRef
+    private Event event;
+    @Field("reg_date")
+    private Date regDate;
+    @Field("reg_number")
+    private Integer regNumber;
+
+    public String getId() {
+        return id;
+    }
+
+    public void setId(String id) {
+        this.id = id;
+    }
+
+    public Event getEvent() {
+        return event;
+    }
+
+    public void setEvent(Event event) {
+        this.event = event;
+    }
+
+    public Date getRegDate() {
+        return regDate;
+    }
+
+    public void setRegDate(Date regDate) {
+        this.regDate = regDate;
+    }
+
+    public Integer getRegNumber() {
+        return regNumber;
+    }
+
+    public void setRegNumber(Integer regNumber) {
+        this.regNumber = regNumber;
+    }
+}
