+++ src/main/java/ru/andnovikov/sportnow/domain/enumeration/EventKind.java
@@ -1,0 +1,5 @@
+package ru.andnovikov.sportnow.domain.enumeration;
+
+public enum EventKind {
+    RUNNING, SWIMMING, TRIATHLON, BIKE, SKIING;
+}
