+++ src/main/java/ru/andnovikov/sportnow/domain/enumeration/EventStatus.java
@@ -1,0 +1,5 @@
+package ru.andnovikov.sportnow.domain.enumeration;
+
+public enum EventStatus {
+    PLAN, OPEN, CLOSED, COMPLETE
+}
