+++ src/main/java/ru/andnovikov/sportnow/repository/UserRepository.java
@@ -1,0 +1,30 @@
+package ru.andnovikov.sportnow.repository;
+
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.mongodb.repository.MongoRepository;
+import org.springframework.stereotype.Repository;
+import ru.andnovikov.sportnow.domain.User;
+
+import java.time.Instant;
+import java.util.List;
+import java.util.Optional;
+
+/**
+ * Spring Data MongoDB repository for the {@link User} entity.
+ */
+@Repository
+public interface UserRepository extends MongoRepository<User, String> {
+
+    Optional<User> findOneByActivationKey(String activationKey);
+
+    List<User> findAllByActivatedIsFalseAndActivationKeyIsNotNullAndCreatedDateBefore(Instant dateTime);
+
+    Optional<User> findOneByResetKey(String resetKey);
+
+    Optional<User> findOneByEmailIgnoreCase(String email);
+
+    Optional<User> findOneByLogin(String login);
+
+    Page<User> findAllByLoginNot(Pageable pageable, String login);
+}
