+++ src/main/resources/application-dev.yml
@@ -1,6 +1,5 @@
 logging:
   level:
-    ROOT: DEBUG
     ru.andnovikov.ijustwannarun: DEBUG
 
 spring:
@@ -10,7 +9,7 @@ spring:
   data:
     mongodb:
       uri: mongodb://localhost:27017
-      database: ijustwannarun
+      database: sportnow
 
   thymeleaf:
     cache: false
