--- src/test/java/ru/andnovikov/ijustwannarun/IJustWannaRunApplicationTests.java
@@ -1,13 +1,0 @@
-package ru.andnovikov.ijustwannarun;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class IJustWannaRunApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
