+++ src/test/java/ru/andnovikov/sportnow/IJustWannaRunApplicationTests.java
@@ -1,0 +1,13 @@
+package ru.andnovikov.sportnow;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class IJustWannaRunApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
